refactor: Home のエントリーポイントの削除
