+++ src/auth/PrivateRoute.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from "react";
 import { Route, RouteProps } from "react-router-dom";
 import { AuthContext } from "./AuthProvider";
 
-import { Home } from "../components";
+import Home from "../components/Home/Home";
 
 /**
  * 認証済みを判断し、認証状態の場合そのまま、非認証状態の場合ホームに移動
